User judgment

+++ app/controllers/items_controller.rb
@@ -37,6 +37,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    redirect_to root_path unless item.user == current_user
     item.destroy
     redirect_to root_path
   end
